Feat/pricing (#9811)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added two new keys (`'guide_flow_modal_contact_sales'`
and `'guide_flow_modal_get_started'`) to the `StringKeysMapType` type in
the `core` package, expanding the available options for localization.
- New Feature: Introduced two new level types (`'starter'` and
`'business'`) to the `LevelType` enum and the `ISpaceLevelType` union
type in the `space_info` package, providing additional subscription
levels for users.
- New Feature: Modified the `LevelCard` component in the `space_info`
package to display buttons for the new level types (`Starter` and
`Business`), enhancing the user interface with expanded functionality.
- Refactor: Updated the `UpgradeSpace` component in the `upgrade_space`
package to accept an optional `hideDetail` prop, utilize TypeScript with
an interface for props, improve logic, and include a commented-out line
for local development URL.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->