fix: fix multple select (#9866)
@coderabbitai:  ignore

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: The `handleSelect` function in the `ListDropdown` component
has been modified. The line `setOpen(false)` has been commented out,
which may affect the behavior of the code.
- New Feature: A new component called `MultipleSelect` has been added to
the `select.stories.tsx` file. It allows multiple selections in a
dropdown and is demonstrated by the `MultipleSelectDemo` function.
- Refactor: The import of the `useRef` function from the `react` package
has been removed from the `time_input/index.tsx` file. This change does
not impact the external interface or behavior of the code.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->