fix: mirror update frozen column (#9907)
- fix: #9579
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Updated the `getCurrentViewBase` function in the `calc.ts`
file to correctly adjust the `frozenColumnCount` property of the `view`
object when `permissionColumns` are changed. This ensures that the
`frozenColumnCount` is adjusted based on field roles obtained from the
`fieldPermissionMap`. The `columns` property of the `view` object is
also updated with the `permissionColumns`. These changes improve the
accuracy and consistency of the displayed data in the Grid and Gantt
views, enhancing the user experience.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->