fix: sync system config with shell (#9967)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->





<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a new target `sys-config-json` to the Makefile,
allowing for the generation of a system config JSON file using a
TypeScript script. This enables synchronization of the system
configuration with the shell.
- New Feature: Introduced a new module `settingConfig` in the
`enterprise/scripts/setting.ts` file, exporting an array of table
configuration objects. This provides additional parameters for fetching
records from datasheets and improves data handling.
- New Feature: Added a new file `system-config-generator.ts` in the
`enterprise/scripts` directory, which includes functions for requesting
data from datasheets, filtering objects, parsing multiple tables, and
generating a JSON file. It enhances the functionality and logic of the
system configuration process.
- Bug Fix: Removed a specific condition in the `rules.ts` file that
caused a rule with the ID `'edition_IS_vika'` to always return `false`.
This fix ensures that the rule can pass if it exists in the
`rulesConfig` array.
- New Feature: Introduced a new target called "wizard" to the Makefile,
running a TypeScript script for updating the system configuration. This
provides a convenient way to sync the configuration with the shell.
- New Feature: Added a new file `wizard-update.ts` in the
`enterprise/scripts` directory, containing code for syncing the system
configuration with a shell. It includes functions for handling data from
multiple tables, filtering, parsing, and generating JSON files. The
changes optimize data retrieval and processing tasks.
- New Feature: Added a new file `system_config.source.json` to the
`vikadata/apitable/packages/core/src/config` directory, containing an
array of table configurations. This enhances the `settingConfig`
variable in `enterprise/scripts/setting.ts` by including the new
configuration.
- New Feature: Introduced a new file `wizard-update.ts` in the
`enterprise/scripts` directory, which retrieves data from a datasheet,
filters and processes the data, and generates a JSON file. It also
handles related records and updates the JSON file accordingly. The
changes focus on improving data retrieval and processing tasks.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->