Fix/fix style schedule (#10032)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->




<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Style: Cosmetic changes were made to various time-related components,
including styling modifications and text replacements. These changes do
not affect the logic or functionality of the code.
- Style: The code formatting configuration file (.prettierrc) was
modified to specify code formatting settings. These changes do not
affect the logic or functionality of the code.
- Refactor: The `DatasheetRecordAlarmRepository` class had empty lines
and unnecessary line breaks removed in method signatures and
implementations. No changes were made to the `deleteRecordAlarmsByIds`
method.
- New Feature: The `ScheduleRuleDisplay` component now uses the `Timing`
component from `@apitable/components` to display schedule timing
information. This change may impact the visual representation and
behavior of the schedule timing display in the UI.
- Style: An import statement for the `useCssColors` hook was added, and
the `OptionItem` component now uses this hook to set the color of the
`CheckOutlined` component. This change may affect the appearance of the
`CheckOutlined` component.
- Style: The `Timing` component now has additional props (`openSearch`,
`searchPlaceholder`) for the `Select` components and a new `Typography`
component for displaying the "by_at" string. The `listStyle` prop was
also added to adjust the width of the dropdown list. These changes do
not affect the logic or functionality of the code.
- Style: Styling changes were made to input elements, select trigger
elements, and list containers in the time-related components. These
changes are cosmetic and do not affect the logic or functionality of the
code.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->