fix: fix issue for timing minute (#10082)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->




<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: The `getUTCOffset` function in the `utils.ts` file now
returns the UTC offset as a string, including the sign (+/-). It also
falls back to using the UTC offset of the current system time if an
error occurs while retrieving the UTC offset for the specified timezone.
- Style: The `Timing` component in the `timing.tsx` file has been
updated to add a `listStyle` prop with a width of '120px' to the
`Trigger` component, affecting its styling.
- Refactor: The `RobotRunHistoryTriggerDetail` component in the
`robot_run_history_item_detail_trigger.tsx` file has been modified to
handle different cases based on the presence of a `timeZone` property
and the emptiness of the `datasheet`. JSX elements are rendered
accordingly, and the `outputSchema` is enriched based on `fields` and
`fieldPermissionMap`.
- Style: The background color of the `StyledSelectTrigger` component in
the `styled.ts` file has been set to `props.theme.color.fc6`.
- Style: The background color of a styled input component in the
`time_input/index.tsx` file has been modified from
`props.theme.color.bgControlsDefault` to `props.theme.color.fc6`.
- Refactor: Changes have been made to the `helper.ts` file, including
importing the `useMemo` function from React, moving the
`TriggerDataSheetMap` import, and modifying the
`getNodeOutputSchemaList` function to filter out triggers of a specific
type and process the remaining triggers.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->