Hotfix/cola (#10264)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

Release Notes:

- Bug Fix: Removed triggering of a specific event related to a
questionnaire being shown after signing in the `WorkspaceRoute`
component.
- Refactor: Modified the `Modal` component by removing the `description`
prop and adding an `onPlay` prop. The `onClose` and `onPlay` functions
are called when the modal is closed and when the video starts playing,
respectively.
- New Feature: Added a new condition to the `goToUpgrade` function in
`upgrade_method.ts`. If the `subscription.product` includes the string
`'appsumo'`, it opens a new window with a specific URL. This allows for
different behavior based on the value of `subscription.product`.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->