feat: workdoc plugin image preview (#10327)
- fix: #9946


<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: The `PreviewImage` component now correctly assigns the value
of `file.token` to `src` when the `src` variable is empty and the
`file.token` includes the string "http". This ensures that the
`file.token` can be used as the image source when the `isSupportImage`
function returns false.
- Documentation: The interface `IPreviewTypeBase` has been modified to
make several properties optional. This may affect the external interface
and behavior of the code.
- New Feature: The `Resizer` component in the workdoc plugin now
includes an image preview feature. Clicking on the image expands it
using the `expandWorkdocImage` function, allowing users to view the
image in detail and delete it if needed.
- Refactor: The `PreviewType` component has been modified to indicate
that the `userInfo` and `previewUrl` props are expected to always have a
value by adding non-null assertions (`!`). This removes the possibility
of these props being `null` or `undefined`.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->