fix: fix active status for button column (#10352)
close https://github.com/vikadata/vikadata/issues/10310
close https://github.com/vikadata/vikadata/issues/9465
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->




<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added `CellButton` component to the `konva_grid` package,
allowing users to interact with cells in a grid.
- Refactor: Optimized and refactored the `FloatUiTooltip` component in
the `tooltip` package for improved performance and code organization.
- Refactor: Modified rendering logic for the `Button` and `WorkDoc`
field types in the `FieldBlock` component in the `expand_record`
package.
- Bug Fix: Fixed an import statement issue in the `CellValue` component
in the `konva_grid` package.
- Refactor: Introduced a constant and updated the `CellHelper` class in
the `cell_helper` module for better code readability.
- Refactor: Made changes to the `ColorGroup` component in the `common`
package, including adjustments to the rendering logic and the addition
of a tooltip.
- Refactor: Updated the `useFloatUiDropdown` function in the `float_ui`
module in the `dropdown` package, modifying its signature and variable
names for clarity.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->