Feat/embed page (#10376)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Updated the text displayed in a span element and the href
attribute of an anchor element in the "Setting" component of the embed
page. The changes now use translated values from `Strings` using the `t`
function.
- Refactor: Modified the `SettingInner` component to display a
description based on the value of `getEnvVariables().IS_AITABLE`.
- Refactor: Swapped the conditions for returning `AI_TABLE_CONFIG` and
`VIKA_CONFIG` in the `getConfig` function of the embed page
configuration. Now, `AI_TABLE_CONFIG` is returned when `IS_AITABLE` is
true.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->