fix: form workdoc click crash (#10436)
- fix: #10415
- fix: #10407
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: The `RowHeadOperation` component in `row_head_operation.tsx`
now updates the fill color of an SVG icon based on the value of the
`isChecked` prop, using the `colors.primaryColor` instead of
`colors.textStaticTertiary`.
- Bug Fix: The crash issue related to a form workdoc click has been
resolved in `index.tsx`. The fix introduces a new variable `nodeId` to
store the element with the ID 'workdocDownload'. If the element is not
found, the function returns null. The `createPortal` function is then
called with the `dropdownMenu` and `nodeId` as arguments.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->