fix: widget build fail (#10490)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Improved internationalization support in the `i18n.ts`
file. The `t` function now includes a new line and introduces a variable
`lang` derived from `getLanguage()`. This allows for more accurate
language translations by replacing hyphens with underscores. The default
language is set based on the value of
`window.__initialization_data__.envVars.IS_APITABLE`, with fallback to
the default language if a translation is not found.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->