fix: close #10488 (#10500)
close #10488
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Refactor: The `ImageBox` component in the `record_card` module has
been updated to import the `useAllowDownloadAttachment` hook and the
`useAppSelector` hook from different modules. It also introduces a new
variable `datasheetId` obtained from the app state. The `allowDownload`
variable is set using the `useAllowDownloadAttachment` hook, and the
`disabledDownload` prop is added to the `DisplayFile` component, which
is set to `!allowDownload`. These changes improve the functionality and
behavior of the `ImageBox` component.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->