chore: update record activity help url (#10605)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Documentation: Updated the URL for the "record_history_help_url" field
in the `strings_ar-SA` file of the `l10n/base` package. The new URL is
"https://help.apitable.com/docs/guide/manual-record-activity". This
change provides users with an updated help resource for understanding
manual record activity in the application.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->

Co-authored-by: xukecheng <xukecheng@aitable.ai>