fix: open workdoc delete cellvalue (#10630)
- fix: #10618
- fix: #10613
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: The `clear` function in the `shortcut_key.ts` module now
checks for the presence of `recordId` and `fieldId` in query parameters
before executing further. If the field type is `FieldType.WorkDoc` and
the cell value is not empty, the function returns without further
execution.
- Style: Changes to the CSS styling of a video element in the
`styles.module.less` file. The `.video-selected` class now adds an
outline and sets the line-height to 0. The `border` property for the
video element has been removed, and the `pointer-events` property for
the `.video-actions` element is now set to "auto".
- Chore: Removal of an unused import statement for the `Button`
component from `@apitable/components` in the `container.tsx` file.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->