fix: workdoc chinese first letter (#10668)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Refactor: Updated import statement for the `Workdoc` component in
`workdoc_editor.tsx` to import from the `enterprise/editor/workdoc`
module instead of a specific file path.
- Refactor: Removed the `handleDOMEvents` property from the
`editorProps` object in `index.tsx` to disable the default behavior for
rebuilding nodes on the `compositionend` event.
- New Feature: Added two new exports, `workdoc` and
`collaboration_editor`, to the codebase in `index.ts`.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->