fix: link select ignore filter (#10670)
- fix: #10659
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Improved the `SearchContentBase` component in
`search_content.tsx` by applying a filter to `cellValue` based on the
field type. If the field type is `Link`, the `cellValue` remains
unchanged; otherwise, it filters out values that are not present in
`foreignRows` or `archivedRecordIds`.
- Bug Fix: Added filtering to the `rows` array in `search_content.tsx`
based on whether the `recordId` is present in `archivedRecordIds`. This
filtering is applied both before and after the search operation.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->